requirements.txt
numpy
scipy
opencv-python
Pillow
timm
torch
torchvision
opend3d
